pyproject.toml
# tells pip what build tool to use to build your package
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "sport"
readme = "README.md"
version = "0.0.1"
authors = [{ name = "Shreyansh Sharma" }]
description = "Sports Object Recognition and Tracking"

## Dependencies
dependencies = [
    "torch", 
    "torchvision", 
    "numpy", 
    "wandb", 
    "datasets", 
    "av",
    "transformers",
    "opencv-python",
    "pillow",
]
requires-python = ">=3.11"

# Entry Points
[project.scripts]
sport = "sport.cli:main"

# For removing pylint import errors
[MASTER]
init-hook = "from pylint.config import find_pylintrc; import os, sys; sys.path.append(os.path.dirname(find_pylintrc()))"

[tool.setuptools]
py-modules = ["sport"]